update-module-names Split sub menu into separate class.
